requirements.txt
pytorch-ignite
pylsd
vietocr
pandas
seaborn
PyYAML
Cython
scikit-learn
tensorboard